refactor: multi qr section
