+++ src/templates/healthcert/parsers/pdtHealthCertV1/index.ts
@@ -1,3 +1,3 @@
 export * from "./parseInfo";
 export * from "./generateMemoSections";
-export * from "./generateResultSection";
+export * from "./generateMultiQrSection";
